+++ docs/examples/jaffle_shop/models/intermediate/sales/orders/int__sales__orders__enriched.sql
@@ -10,6 +10,7 @@ WITH
 			stg__customers__profiles__clean.customer_first_name,
 			stg__sales__orders__standardized.customer_id,
 			stg__customers__profiles__clean.customer_name,
+			stg__sales__orders__standardized.datetime,
 			stg__sales__orders__standardized.order_date,
 			stg__sales__orders__standardized.order_id,
 			stg__sales__orders__standardized.order_total_cents,
@@ -22,11 +23,18 @@ WITH
 			stg__sales__stores__locations.store_name,
 			stg__sales__stores__locations.store_tax_rate,
 			stg__sales__orders__standardized.subtotal_cents,
-			stg__sales__orders__standardized.tax_paid_cents
+			stg__sales__orders__standardized.tax_paid_cents,
+			-- partition columns
+			stg__sales__orders__standardized.portal_partition_monthly,
+			stg__sales__orders__standardized.portal_partition_daily,
+			stg__sales__orders__standardized.portal_partition_hourly
 		FROM
 			{{ ref('stg__sales__orders__standardized') }} stg__sales__orders__standardized
 			LEFT JOIN {{ ref('stg__customers__profiles__clean') }} stg__customers__profiles__clean ON stg__sales__orders__standardized.customer_id = stg__customers__profiles__clean.customer_id
 			LEFT JOIN {{ ref('stg__sales__stores__locations') }} stg__sales__stores__locations ON stg__sales__orders__standardized.store_id = stg__sales__stores__locations.store_id
+		WHERE
+			{{ _ext_event_date_filter("stg__sales__orders__standardized.portal_partition_monthly", data_type="date", interval="month") }}
+			AND {{ _ext_event_date_filter("stg__sales__orders__standardized.portal_partition_daily", data_type="date") }}
 	)
 SELECT
 	*