+++ docs/examples/jaffle_shop/models/intermediate/sales/orders/int__sales__orders__regional_combined.sql
@@ -10,6 +10,7 @@ WITH
 			customer_first_name,
 			customer_id,
 			customer_name,
+			datetime,
 			order_date,
 			order_id,
 			order_total_cents,
@@ -25,14 +26,21 @@ WITH
 			store_name,
 			store_tax_rate,
 			subtotal_cents,
-			tax_paid_cents
+			tax_paid_cents,
+			portal_partition_monthly,
+			portal_partition_daily,
+			portal_partition_hourly
 		FROM
 			{{ ref('int__sales__orders__enriched') }}
+		WHERE
+			{{ _ext_event_date_filter("portal_partition_monthly", data_type="date", interval="month") }}
+			AND {{ _ext_event_date_filter("portal_partition_daily", data_type="date") }}
 		UNION ALL
 		SELECT
 			customer_first_name,
 			customer_id,
 			customer_name,
+			datetime,
 			order_date,
 			order_id,
 			order_total_cents,
@@ -48,9 +56,15 @@ WITH
 			store_name,
 			store_tax_rate,
 			subtotal_cents,
-			tax_paid_cents
+			tax_paid_cents,
+			portal_partition_monthly,
+			portal_partition_daily,
+			portal_partition_hourly
 		FROM
 			{{ ref('int__sales__orders__enriched') }}
+		WHERE
+			{{ _ext_event_date_filter("portal_partition_monthly", data_type="date", interval="month") }}
+			AND {{ _ext_event_date_filter("portal_partition_daily", data_type="date") }}
 	)
 SELECT
 	*